+++ test/libsolidity/smtCheckerTests/functions/getters/mapping_2.sol
@@ -1,0 +1,13 @@
+pragma experimental SMTChecker;
+
+contract C {
+	mapping (uint => mapping (uint => uint)) public map;
+
+	function f() public view {
+		uint y = this.map(2, 3);
+		assert(y == map[2][3]); // should hold
+		assert(y == 1); // should fail
+	}
+}
+// ----
+// Warning 6328: (199-213): CHC: Assertion violation happens here.