+++ test/libsolidity/smtCheckerTests/functions/getters/string.sol
@@ -1,0 +1,13 @@
+pragma experimental SMTChecker;
+
+contract C {
+	string public str1 = 'b';
+
+	function f() public view {
+		string memory a1 = this.str1();
+		assert(keccak256(bytes(a1)) == keccak256(bytes(str1))); // should hold
+		assert(keccak256(bytes(a1)) == keccak256('a')); // should fail
+	}
+}
+// ----
+// Warning 6328: (211-257): CHC: Assertion violation happens here.