+++ test/libsolidity/smtCheckerTests/operators/bitwise_not_fixed_bytes.sol
@@ -1,7 +1,12 @@
 pragma experimental SMTChecker;
+
 contract C {
-  function f() public pure returns (byte) {
-    return (~byte(0xFF));
-  }
+	function f() public pure {
+		// ffff0000 in bytes4
+		bytes4 x = ~bytes4(hex"ffff");
+		assert(x == 0xffff0000); // should fail
+		assert(x == 0x0000ffff); // should hold
+	}
 }
 // ----
+// Warning 6328: (175-198): CHC: Assertion violation happens here.